requirements.txt
# Google Ads API
google-ads==25.1.0
google-auth-oauthlib==1.2.1

# Meta Marketing API
facebook-business==21.0.5

# LinkedIn Ads (não tem SDK oficial maintido; usamos requests)
requests==2.32.3

# Utils
python-dotenv==1.0.1
pandas==2.2.3
click==8.1.7
rich==13.9.4
pyyaml==6.0.2
